+++ CHANGELOG.md
@@ -48,6 +48,8 @@ Changes before Tatum release are not documented in this file.
 
 #### Fixed
 
+- Signature output in `subscribe --with-metadata` (https://github.com/streamr-dev/network/pull/3245)
+
 #### Security